Split graph and tree attributes (code)
