+++ higra/attribute/all.hpp
@@ -10,5 +10,5 @@
 
 #pragma once
 
-#include "py_attributes.hpp"
+#include "py_tree_attributes.hpp"
